Improve the MSBuild task building and packaging for full .NET
The full framework console app did not reference the NS2 MSBuild
task, instead having its own copy of the task class. This was unnecessary
since the task class already lives in a NS2 library.

Also, the .props file from the package was not referencing the MSBuild
assembly from the location of the desktop assemblies, instead loading
the netcore path version. This was risky since the wrong assemblies
could have been loaded inadvertently.

This commit improves that by conditionally importing the tasks
assembly from either the desktop or the netcore locations, which
would then cause any other dependencies in that folder to be resolved
for the right framework.